+++ inst/examples/ex-curvefits.R
@@ -1,5 +1,5 @@
 # Rough fitting and growing season dividing
-brks2 <- season_3y(INPUT,
+brks2 <- season_mov(INPUT,
     rFUN = wWHIT, wFUN = wFUN,
     plotdat = d, IsPlot = IsPlot, print = FALSE, IsPlot.OnlyBad = FALSE)
 # Fine fitting