+++ inst/examples/ex-season.R
@@ -4,7 +4,7 @@ brks  <- season(INPUT,
     lambda = 10,
     plotdat = d, IsPlot = IsPlot, print = FALSE, IsPlot.OnlyBad = FALSE)
 # curve fitting by year
-brks2 <- season_3y(INPUT,
+brks2 <- season_mov(INPUT,
     rFUN = wWHIT, wFUN = wFUN,
     lambda = 10,
     plotdat = d, IsPlot = IsPlot, print = FALSE, IsPlot.OnlyBad = FALSE)