+++ inst/shiny/check_season/ui.R
@@ -16,7 +16,7 @@ ui <- fluidPage(
             titlePanel("Phenofit: growing season dividing"),
             numericInput("iters", "iters:", 2, 1, 3),
             selectInput("FUN_season","Choose a season dividing function (FUN_season):",
-                        choices = c('season', 'season_3y'), selected = "season_3y"),
+                        choices = c('season', 'season_mov'), selected = "season_mov"),
             selectInput("wFUN","Choose a weights updating function (wFUN):",
                         choices = c("wTSM", "wBisquare", "wChen"), selected = "wBisquare"),
             selectInput("rFUN","Choose a Rough fitting function (FUN):",
@@ -39,7 +39,7 @@ ui <- fluidPage(
                 numericInput("nf", "number of frequencies (nf):", 3, 1, 6)
             ),
             conditionalPanel(
-                condition = "input.FUN_season == 'season_3y'",
+                condition = "input.FUN_season == 'season_mov'",
                 numericInput("maxExtendMonth", "Include n previous and subsequent month (maxExtendMonth):",
                     2, 0, 12)
             ),